+++ components/Auth/ProfilePage/index.tsx
@@ -1,6 +1,6 @@
 import { memo, useCallback, useState } from 'react'
 import Avatar from '@elements/Avatar'
-import { User } from '@helpers/types'
+import { User } from '@lib/types'
 import EditDetails from './EditDetails'
 import Socials from './Socials'
