+++ components/Dashboard/AdminDashboardPage/UsersTable/index.tsx
@@ -2,8 +2,8 @@ import { memo, useCallback, useState } from 'react'
 import dynamic from 'next/dynamic'
 import useSWR from 'swr'
 import Avatar from '@elements/Avatar'
-import { Role, User } from '@helpers/types'
-import { useUser } from '@helpers/user'
+import { Role, User } from '@lib/types'
+import { useUser } from '@lib/user'
 import RoleTags from '../RoleTags'
 import SearchUser from './SearchUser'
 const DeleteUserModal = dynamic(() => import('./DeleteUserModal'))
