+++ components/Events/EventPage/index.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import Link from 'next/link'
-import { ImageProps } from '@helpers/types'
+import { ImageProps } from '@lib/types'
 
 const EventPage = ({ data }: Props) => (
   <>
