+++ pages/api/announcements.ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import dayjs from 'dayjs'
 import dbConnect from '@lib/dbConnect'
 import announcements from '@models/announcements'
-import { Announcements } from '@helpers/types'
+import { Announcements } from '@lib/types'
 
 const announcementRoute = async (req: NextApiRequest, res: NextApiResponse) => {
   const { method, body } = req
