+++ pages/api/users/[id].ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import dayjs from 'dayjs'
 import User from '@models/user'
 import dbConnect from '@lib/dbConnect'
-import { User as UserType } from '@helpers/types'
+import { User as UserType } from '@lib/types'
 
 const convertDate = (date: string) => dayjs(date).format('MMM D, YYYY h:mm A')
